Try a new stylistic approach
Float all the letters to the left seems to look pretty righteous. Also,
I've started working on this thing called Quicklook (tm). Quicklook
opens a blog post without changing the page. Right now all it does is
slide a poster to the upper left of the window, but give it time, and it
might ferment into something neat.

Quicklook is disabled right now, just in case we happen to make a real
post at some point before it's done. To enable it, just change
SETTINGS.QUICKLOOK to true in index.haml